Added rook cleanup helper
This helper is for cleaning up a single node with multiple devices. It
should be deployed as multiple aliases in helm dependencies to wipe
multiple nodes. It will not do anything by itself, and requires manual
triggering as it deploys a suspended cronjob that required manual
activation since it is only useful in disaster scenarios to clean up.

Signed-off-by: GeorgeRaven <GeorgeRavenCommunity@pm.me>