Added rollout strategy override and automation
Rollout strategy for deployments is now volume and autoscaler aware.
It will now default to `Recreate` when one or more volumes and a replica of 1 is
found with no autoscaler. This ensures rollouts can still occur when
there is only one replica which would otherwise stall.

Improvement would be to also add volume permission detection, so this
only happens when RWO is the only setting. Since RWX would still work
even with non Recreate strategy.