Added most basic postgresql chart
This is intended to be used as a battery for battery-included charts so
users can quickly test an application. The intention is that this
postgresql chart is disabled in production, and instead the user would
use something like CNPG. Thus I am not too fussed about this chart, it
is only here to facilitate other charts to be tested
etc.